pyproject.toml
# pyproject.toml — v3
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ayextractor"
version = "0.3.0"
description = "Multi-Agent Document Analyzer with GraphRAG"
requires-python = ">=3.11"
license = {text = "Proprietary"}

dependencies = [
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.0",
    "numpy>=2.4.2",
    "networkx>=3.6.1",
    "scipy>=1.17.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.1",
    "mypy>=1.19.1",
]
pdf = [
    "pymupdf>=1.27.1",
]
llm = [
    "anthropic>=0.79.0",
    "openai>=2.21.0",
    "google-generativeai>=0.8.6",
    "ollama>=0.6.1",
]
graph = [
    "leidenalg>=0.11.0",
    "python-igraph>=1.0.0",
]
rag = [
    "chromadb>=1.5.0",
    "langchain-core>=1.2.13",
    "langgraph>=1.0.8",
]
gpu = [
    "nx-cugraph-cu12>=26.2.0",
    "cuml-cu12>=26.2.0",
]
all = ["ayextractor[pdf,llm,graph,rag]"]

[tool.setuptools.package-dir]
ayextractor = "src"

[tool.setuptools]
packages = [
    "ayextractor",
    "ayextractor.api",
    "ayextractor.batch",
    "ayextractor.cache",
    "ayextractor.chunking",
    "ayextractor.config",
    "ayextractor.consolidator",
    "ayextractor.core",
    "ayextractor.extraction",
    "ayextractor.graph",
    "ayextractor.graph.layers",
    "ayextractor.graph.profiles",
    "ayextractor.llm",
    "ayextractor.llm.adapters",
    "ayextractor.logging",
    "ayextractor.pipeline",
    "ayextractor.pipeline.agents",
    "ayextractor.pipeline.plugin_kit",
    "ayextractor.pipeline.prompts",
    "ayextractor.rag",
    "ayextractor.rag.embeddings",
    "ayextractor.rag.graph_store",
    "ayextractor.rag.retriever",
    "ayextractor.rag.vector_store",
    "ayextractor.storage",
    "ayextractor.tracking",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "gpu: GPU-accelerated tests (requires NVIDIA GPU + RAPIDS)",
    "integration: integration tests requiring external services",
    "slow: long-running tests",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true